Moving hooks from utils to hooks/hooks.ts
